Improve loading of fonts
While it worked with the previous method, it had several drawbacks :
* highend users had the disgraceful FOUT
* the font was never loaded for js-less users

This new method solves it by displaying the fallback only when we
detect that the fonts are slow to load (> 1s). If 5s after load the fonts
are still not loaded, we ignore the main font and keep the fallback font.
The goal here is to avoid content jumping since the user is likely to be
reading something already.